[#265] Quality: Changed package name
