+++ app/src/androidTest/java/com/example/bigowlapp/activity/HomePageActivityTest.java
@@ -9,7 +9,7 @@ import com.example.bigowlapp.R;
 import com.example.bigowlapp.model.LiveDataWithStatus;
 import com.example.bigowlapp.model.User;
 import com.example.bigowlapp.utils.MemberScheduleAlarmManager;
-import com.example.bigowlapp.viewModel.HomePageViewModel;
+import com.example.bigowlapp.view_model.HomePageViewModel;
 
 import org.junit.Before;
 import org.junit.Rule;
