+++ app/src/androidTest/java/com/example/bigowlapp/activity/ScheduleReportActivityTest.java
@@ -18,7 +18,7 @@ import com.example.bigowlapp.model.Response;
 import com.example.bigowlapp.model.Schedule;
 import com.example.bigowlapp.model.User;
 import com.example.bigowlapp.model.UserScheduleResponse;
-import com.example.bigowlapp.viewModel.ScheduleReportViewModel;
+import com.example.bigowlapp.view_model.ScheduleReportViewModel;
 import com.google.firebase.Timestamp;
 import com.google.firebase.firestore.GeoPoint;
