+++ app/src/main/java/com/example/bigowlapp/view_model/BaseViewModel.java
@@ -1,4 +1,4 @@
-package com.example.bigowlapp.viewModel;
+package com.example.bigowlapp.view_model;
 
 import androidx.lifecycle.ViewModel;
