+++ app/src/main/java/com/example/bigowlapp/view_model/LogInViewModel.java
@@ -1,14 +1,21 @@
-package com.example.bigowlapp.viewModel;
+package com.example.bigowlapp.view_model;
 
 import com.example.bigowlapp.model.LiveDataWithStatus;
 import com.example.bigowlapp.model.User;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
 
-public class HomePageViewModel extends BaseViewModel {
+public class LogInViewModel extends BaseViewModel {
 
     private LiveDataWithStatus<User> user;
 
-    public HomePageViewModel() {
+    public LogInViewModel() {
         // used implicitly when ViewModel constructed using ViewModelProvider
+    }
+
+    public Task<AuthResult> logInUser(String email, String password) {
+        return repositoryFacade.getAuthRepository().signInUser(email, password);
     }
 
     public LiveDataWithStatus<User> getCurrentUserData() {
@@ -19,8 +26,8 @@ public class HomePageViewModel extends BaseViewModel {
         return user;
     }
 
-    public void updateUser(User user){
-        repositoryFacade.getUserRepository().updateDocument(user.getUid(), user);
+    public void addAuthStateListenerToDatabase(FirebaseAuth.AuthStateListener authStateListener) {
+        repositoryFacade.getAuthRepository().addAuthStateListener(authStateListener);
     }
 
     private void loadUserCurrentProfile() {
@@ -28,7 +35,4 @@ public class HomePageViewModel extends BaseViewModel {
                 .getDocumentByUid(getCurrentUserUid(), User.class);
     }
 
-    public void signOut() {
-        repositoryFacade.getAuthRepository().signOutUser();
-    }
 }
