+++ app/src/main/java/com/example/bigowlapp/view_model/SignUpViewModel.java
@@ -1,4 +1,4 @@
-package com.example.bigowlapp.viewModel;
+package com.example.bigowlapp.view_model;
 
 import android.content.Context;
